server/gateway/requirements.txt
cbor2>=5.4
paho-mqtt>=1.6,<2.0
